refactoring mul to MulWordsGadget
